LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/relaxing-the-same-origin-restriction/document_domain_access_details.sub-expected.txt



Harness Error (TIMEOUT), message = null

PASS Access allowed if same-origin with no 'document.domain' modification. (Sanity check)
PASS Access not allowed if different-origin with no 'document.domain' modification. (Sanity check)
PASS Access disallowed if same-origin but only one sets document.domain.
PASS Access allowed if same-origin and both set document.domain to existing value.
PASS Access allowed if different-origin but both set document.domain to parent domain.
PASS Access disallowed again if same-origin, both set document-domain to existing value, then one sets to parent.
PASS Access is revoked to Window object when we stop being same effective script origin due to document.domain.
FAIL Access is revoked to Location object when we stop being same effective script origin due to document.domain. assert_equals: expected "SecurityError" but got "Reachable 3"
PASS Access is not revoked to Document object when we stop being same effective script origin due to document.domain.
PASS Access is not revoked to random object when we stop being same effective script origin due to document.domain.
TIMEOUT Access evolves correctly for non-cross-origin objects when we join up via document.domain and then diverge again. Test timed out
NOTRUN Access evolves correctly for cross-origin objects when we join up via document.domain and then diverge again.
